Gallery
- Reworked the redux store to the service pattern
- Separated the store actions service from the enums; added type guards and types to the enums
- Fixed a bunch of tests suites
- Implemented layout for Gallery columns
- Changed graphs to be more reactive to changes
- Implemented most of the Gallery UI and backend read operations
- Removed the Identifiable interface

+++ api-server/database/budgetItem.dao.ts
@@ -63,7 +63,7 @@ export class BudgetItem extends Model<BudgetItemAttributes, BudgetItemCreationAt
     public description!: string | null;
 
     @AllowNull(false)
-    @Column(DataType.INTEGER)
+    @Column(DataType.BIGINT)
     public totalPrice!: number;
 
     @AllowNull(false)
